+++ src/util/timetaker.h
@@ -30,7 +30,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 class TimeTaker
 {
 public:
-	TimeTaker(const char *name, u32 *result=NULL,
+	TimeTaker(const char *name, u64 *result=NULL,
 		TimePrecision=PRECISION_MILLI);
 
 	~TimeTaker()
@@ -38,16 +38,16 @@ public:
 		stop();
 	}
 
-	u32 stop(bool quiet=false);
+	u64 stop(bool quiet=false);
 
-	u32 getTimerTime();
+	u64 getTimerTime();
 
 private:
 	const char *m_name;
-	u32 m_time1;
+	u64 m_time1;
 	bool m_running;
 	TimePrecision m_precision;
-	u32 *m_result;
+	u64 *m_result;
 };
 
 #endif
